+++ R/accuracy.R
@@ -1,5 +1,5 @@
 ## - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - |
-#     Copyright (C) 2020 - 2021  Reza Mohammadi & Kevin Burke                  |
+#     Copyright (C) 2020  Reza Mohammadi & Kevin Burke                         |
 #                                                                              |
 #     This file is part of 'liver' package.                                    |
 #                                                                              |
@@ -9,20 +9,18 @@
 #                                                                              |
 #     Maintainer: Reza Mohammadi <a.mohammadi@uva.nl>                          |
 ## - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - |
-#     Plot a Confusion Matrix
+#     Compute average classification accuracy
 ## - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - |
 
-conf.mat.plot = function( pred, actual, cutoff = NULL, reference = NULL, 
-                          conf.level = 0, margin = 1, 
-                          color = c( "#ff83a8", "#83ff9b" ), ... )
+accuracy = function( pred, actual, cutoff = NULL, reference = NULL )
 {
     conf_mat = liver::conf.mat( pred = pred, actual = actual, cutoff = cutoff, reference = reference )
     
-    graphics::fourfoldplot( conf_mat, conf.level = conf.level, 
-                            margin = margin, color = color, ... )
-}   
+    accuracy_value = sum( diag( conf_mat ) ) / sum( conf_mat )
+    
+    return( accuracy_value )    
+}
+   
 ## - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - |
-
-