+++ R/kNN.plot.R
@@ -12,7 +12,8 @@
 #     k-Nearest Neighbour Classification using formula interface
 ## - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - |
 
-kNN.plot = function( formula, train, test, k.max = 10, transform = FALSE, set.seed = NULL, ... ) 
+kNN.plot = function( formula, train, test, k.max = 10, transform = FALSE, 
+                     base = "error", set.seed = NULL, ... ) 
 {
     if( k.max < 2 ) stop( "  k.max must be > 1" )
     
@@ -20,15 +21,17 @@ kNN.plot = function( formula, train, test, k.max = 10, transform = FALSE, set.se
     
     formula = stats::as.formula( formula )
     
-    model_frame_train  = stats::model.frame( formula, data = train )
-    model_train        = attr( model_frame_train, "terms" )
-    model_train_no_res = stats::delete.response( model_train )
+    model_frame_train = stats::model.frame( formula, data = train )
+    model_train       = attr( model_frame_train, "terms" )
     
     train_label = stats::model.response( model_frame_train )
-    test_label  = stats::model.matrix( model_train_no_res, test )[ , 1 ]
+    train       = stats::model.matrix( model_train, model_frame_train )[ , -1 ]
     
-    train = stats::model.matrix( model_train, model_frame_train )[ , -1 ]
-    test  = stats::model.matrix( model_train_no_res, test )[ , -1 ]
+    model_frame_test = stats::model.frame( formula, data = test )
+    model_test       = attr( model_frame_test, "terms" )
+    
+    test_label = stats::model.response( model_frame_test )
+    test       = stats::model.matrix( model_test, model_frame_test )[ , -1 ]
     
     if( transform == TRUE ) transform = "minmax"
     
@@ -50,28 +53,51 @@ kNN.plot = function( formula, train, test, k.max = 10, transform = FALSE, set.se
     }
     
     k_list   = 1:k.max
-    mse_list = vector( length = k.max  )
+    base_list = vector( length = k.max  )
     
     for( k in k_list ){
-        knn_k = class::knn( train = train, test = test, cl = train_label, k = k, ... )
+        #knn_k = class::knn( train = train, test = test, cl = train_label, k = k, ... )
+        knn_k = class::knn( train = train, test = test, cl = train_label, k = k )
         
-        mse_list[ k ] = liver::mse( knn_k, test_label )
+        base_list[ k ] = liver::accuracy( knn_k, test_label )
     }
     
-    df = data.frame( k_list = as.factor( k_list ), mse_list = mse_list, stringsAsFactors = TRUE)
+    if( ( base == "accuracy" ) | ( base == "Accuracy" ) )
+    {
+        title = "Optimal value of k based on accuracy"
+        y_lab = "Accuracy"
+    }
     
-    df_gg = data.frame( k_list = k_list, mse_list = mse_list )
+    if( ( base == "MSE" ) | ( base == "mse" ) )
+    {
+        base_list = 1 - base_list
+        
+        title = "Optimal value of k based on MSE"
+        y_lab = "Mean Square Error (MSE)"
+    }
     
-    ggplot2::ggplot( df_gg, ggplot2::aes( x = k_list, y = mse_list ) ) +
+    if( ( base == "Error" ) | ( base == "error" ) )
+    {
+        base_list = 1 - base_list
+        
+        title = "Optimal value of k based on Error Rate"
+        y_lab = "Error Rate"
+    }
+    
+    df = data.frame( k_list = as.factor( k_list ), base_list = base_list, stringsAsFactors = TRUE)
+    
+    df_gg = data.frame( k_list = k_list, base_list = base_list )
+    
+    ggplot2::ggplot( df_gg, ggplot2::aes( x = k_list, y = base_list ) ) +
         ggplot2::geom_line( color = "#a0a0a0" ) + 
         ggplot2::geom_point( shape = 21, color = "#ff5085", fill = "#ff83a8", size = 2 ) + 
         ggplot2::theme_minimal() + 
         ggplot2::scale_x_continuous( breaks = k_list ) +
-        ggplot2::ggtitle( "Optimal value of k based on MSE" ) +
-        ggplot2::labs( x = "Value of k", y = "Mean Square Error (MSE)" ) +
+        ggplot2::ggtitle( title ) +
+        ggplot2::labs( x = "Value of k", y = y_lab ) +
         ggplot2::theme( axis.line = ggplot2::element_line( size = 0.4, colour = "black" ),
-               panel.grid.major = ggplot2::element_line( colour = "#f2f2f2" ), panel.grid.minor = ggplot2::element_blank(),
-               panel.border = ggplot2::element_blank(), panel.background = ggplot2::element_blank() )
+                        panel.grid.major = ggplot2::element_line( colour = "#f2f2f2" ), panel.grid.minor = ggplot2::element_blank(),
+                        panel.border = ggplot2::element_blank(), panel.background = ggplot2::element_blank() )
 }
    
 ## - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - |